Code refactor and fix spacing
