+++ src/components/FilterPresetsSelect/messages.ts
@@ -1,0 +1,12 @@
+import { defineMessages } from "react-intl";
+
+export const messages = defineMessages({
+  unsavedPreset: {
+    defaultMessage: "Unsaved preset",
+    id: "A+g/VP",
+  },
+  filterPreset: {
+    defaultMessage: "Filter preset",
+    id: "3PVGWj",
+  },
+});
